refactor: Extract WithVariable trait
- Extract WithVariable trait to be used by Parameter and Attribute
