+++ src/Code/Variables/WithVariable.php
@@ -1,0 +1,36 @@
+<?php declare(strict_types=1);
+/**
+ * PHP version 7.2
+ *
+ * This source file is subject to the license that is bundled with this package in the file LICENSE.
+ */
+
+namespace PhUml\Code\Variables;
+
+use PhUml\Code\Name;
+
+trait WithVariable
+{
+    /** @var Variable */
+    private $variable;
+
+    public function isAReference(): bool
+    {
+        return $this->variable->isAReference();
+    }
+
+    public function referenceName(): Name
+    {
+        return $this->variable->referenceName();
+    }
+
+    public function hasTypeDeclaration(): bool
+    {
+        return $this->variable->hasTypeDeclaration();
+    }
+
+    public function type(): TypeDeclaration
+    {
+        return $this->variable->type();
+    }
+}
